returns animals back into existance
