+++ simplerogue/src/main/java/game/gameobjects/entities/ThrownItem.java
@@ -11,13 +11,14 @@ import game.gamelogic.Aimable;
 import game.gamelogic.Examinable;
 import game.gamelogic.HasDodge;
 import game.gamelogic.HasResistances;
+import game.gamelogic.behavior.Behavable;
 import game.gamelogic.resistances.PercentageResistance;
 import game.gamelogic.resistances.Resistance;
 import game.gameobjects.DamageType;
 import game.gameobjects.DisplayableTile;
 import game.gameobjects.Space;
 
-public class ThrownItem extends Entity implements HasDodge, HasResistances{
+public class ThrownItem extends Entity implements HasDodge, HasResistances, Behavable{
 
     private Iterator<Space> pathToTarget;
     private Aimable aimable;
